Update about-app-v-51.md
In line 106, I removed the section that states you cannot upgrade directly from v4.x to v5.1.  This is incorrect.  There is certainly a pacakge converter built into the v5.1 sequencer, which is in fact better then the converter in the 5.0 versions of the product.  A v5.0 client doesn't need to be installed prior to installing the v5.1 client.  Upgrading the sequencer isn't supported, you will be building a new one.  Upgrading the server infrastructure isn't suppported, you will be building new servers.  And, as you have noted beginning in line 202, the package format is the same for all versions of 5.x.  So, why would there be a restriction on upgrading from v 4.x directly to version 5.1?